refactor: move literal wire carrier into graphforge-value
Keep IrLiteral as a re-export and preserve the existing custom serialized variants and non-finite float encoding. Validation: cargo test -p graphforge-ir --lib expr::tests (18 passed). #1012 remains in progress for heterogeneous Arrow codecs and durable acceptance.
